Rewrite app structure
