+++ rasp.ap1/rasp.ap1Requests.js
@@ -4,6 +4,4 @@ var loadPage = function (uri, options) {
   return request.get(uri, { timeout: options.time })
 };
 
-module.exports = {
-  loadPage: loadPage
-};
+module.exports = loadPage;
