+++ test/test.js
@@ -1,6 +1,6 @@
 var request = require('request');
 var expect = require('chai').expect;
-var fetchCityBusStops = require('../osm/osm').fetchCityBusStops;
+var fetchCityBusStops = require('../osm/osmRequests').fetchCityBusStops;
 var transformToGeoJson = require('../osm/transformNodesToGeoJson')();
 var stream = require('stream').Stream;
 var config = require('../config').get('requestOptions');
